refactor(atelier_core): split codegen/slots into cohesive modules
