+++ crates/vize_atelier_core/src/codegen/slots/generate.rs
@@ -1,517 +1,19 @@
-//! Slot generation functions.
-//!
-//! Generates slot objects for component children.
+//! Slots object generation for component children.
 
 use crate::ast::*;
 use crate::transforms::v_slot::{collect_slots, get_slot_name, has_v_slot};
-
-use super::context::CodegenContext;
-use super::expression::generate_expression;
-use super::helpers::{camelize, escape_js_string, is_valid_js_identifier};
-use super::node::generate_node;
-use super::props::{
-    generate_slot_outlet_directive_prop_with_static, generate_vbind_object_exp,
-    generate_von_object_exp, is_supported_directive,
-};
 use vize_carton::String;
 use vize_carton::ToCompactString;
 
-pub(crate) enum SlotOutletName<'a> {
-    Static(String),
-    Dynamic(&'a ExpressionNode<'a>),
-}
-
-/// Get slot props expression as raw source (not transformed)
-fn get_slot_props(dir: &DirectiveNode<'_>) -> Option<vize_carton::String> {
-    dir.exp.as_ref().map(|exp| match exp {
-        ExpressionNode::Simple(s) => s.loc.source.clone(),
-        ExpressionNode::Compound(c) => c.loc.source.clone(),
-    })
-}
-
-/// Add _ctx. prefix to default value identifiers in destructuring patterns.
-/// e.g., "{ item = defaultItem }" -> "{ item = _ctx.defaultItem }"
-/// Only processes identifiers after `=` (default values), not the param names.
-fn prefix_slot_defaults(source: &str) -> String {
-    let bytes = source.as_bytes();
-    let len = bytes.len();
-    let mut result = String::with_capacity(len + 20);
-    let mut i = 0;
-
-    while i < len {
-        if bytes[i] == b'=' {
-            // Skip == and =>
-            if i + 1 < len && (bytes[i + 1] == b'=' || bytes[i + 1] == b'>') {
-                result.push(bytes[i] as char);
-                result.push(bytes[i + 1] as char);
-                i += 2;
-                continue;
-            }
-            result.push('=');
-            i += 1;
-            // Skip whitespace after =
-            while i < len && (bytes[i] == b' ' || bytes[i] == b'\t') {
-                result.push(bytes[i] as char);
-                i += 1;
-            }
-            // Check if next is a simple identifier (not a literal/number/string/object)
-            if i < len && (bytes[i].is_ascii_alphabetic() || bytes[i] == b'_' || bytes[i] == b'$') {
-                // Collect the identifier
-                let start = i;
-                while i < len
-                    && (bytes[i].is_ascii_alphanumeric() || bytes[i] == b'_' || bytes[i] == b'$')
-                {
-                    i += 1;
-                }
-                let ident = &source[start..i];
-                // Don't prefix keywords/literals
-                if !matches!(
-                    ident,
-                    "true" | "false" | "null" | "undefined" | "NaN" | "Infinity"
-                ) {
-                    result.push_str("_ctx.");
-                }
-                result.push_str(ident);
-            }
-        } else {
-            result.push(bytes[i] as char);
-            i += 1;
-        }
-    }
-
-    result
-}
-
-/// Extract parameter names from slot props expression
-/// e.g., "{ item }" -> ["item"], "{ item, index }" -> ["item", "index"]
-/// e.g., "slotProps" -> ["slotProps"]
-fn extract_slot_params(props_str: &str) -> Vec<String> {
-    let mut params = Vec::new();
-    super::v_for::extract_destructure_params(props_str.trim(), &mut params);
-    params
-}
-
-fn is_slot_name_bind(dir: &DirectiveNode<'_>) -> bool {
-    if dir.name.as_str() != "bind" {
-        return false;
-    }
-
-    match dir.arg.as_ref() {
-        Some(ExpressionNode::Simple(exp)) => exp.is_static && exp.content.as_str() == "name",
-        _ => false,
-    }
-}
-
-fn is_slot_name_prop(prop: &PropNode<'_>) -> bool {
-    match prop {
-        PropNode::Attribute(attr) => attr.name.as_str() == "name",
-        PropNode::Directive(dir) => is_slot_name_bind(dir),
-    }
-}
-
-fn is_slot_outlet_object_spread(prop: &PropNode<'_>) -> bool {
-    matches!(
-        prop,
-        PropNode::Directive(dir)
-            if (dir.name.as_str() == "bind" || dir.name.as_str() == "on")
-                && dir.arg.is_none()
-                && dir.exp.is_some()
-    )
-}
-
-fn slot_outlet_prop_generates_output(prop: &PropNode<'_>) -> bool {
-    if is_slot_name_prop(prop) {
-        return false;
-    }
-
-    match prop {
-        PropNode::Attribute(_) => true,
-        PropNode::Directive(dir) => {
-            if (dir.name.as_str() == "bind" || dir.name.as_str() == "on") && dir.arg.is_none() {
-                return dir.exp.is_some();
-            }
-            is_supported_directive(dir)
-        }
-    }
-}
-
-fn has_slot_outlet_vbind_object(el: &ElementNode<'_>) -> bool {
-    el.props.iter().any(|prop| {
-        matches!(
-            prop,
-            PropNode::Directive(dir)
-                if dir.name.as_str() == "bind" && dir.arg.is_none() && dir.exp.is_some()
-        )
-    })
-}
-
-fn has_slot_outlet_von_object(el: &ElementNode<'_>) -> bool {
-    el.props.iter().any(|prop| {
-        matches!(
-            prop,
-            PropNode::Directive(dir)
-                if dir.name.as_str() == "on" && dir.arg.is_none() && dir.exp.is_some()
-        )
-    })
-}
-
-fn has_slot_outlet_entry_props(el: &ElementNode<'_>) -> bool {
-    el.props
-        .iter()
-        .any(|prop| !is_slot_outlet_object_spread(prop) && slot_outlet_prop_generates_output(prop))
-}
-
-pub(crate) fn get_slot_outlet_name<'a>(el: &'a ElementNode<'a>) -> SlotOutletName<'a> {
-    for prop in &el.props {
-        match prop {
-            PropNode::Attribute(attr) if attr.name.as_str() == "name" => {
-                let name = attr
-                    .value
-                    .as_ref()
-                    .map(|v| v.content.clone())
-                    .unwrap_or_else(|| String::new("default"));
-                return SlotOutletName::Static(name);
-            }
-            PropNode::Directive(dir) if is_slot_name_bind(dir) => {
-                if let Some(exp) = dir.exp.as_ref() {
-                    return SlotOutletName::Dynamic(exp);
-                }
-            }
-            _ => {}
-        }
-    }
-
-    SlotOutletName::Static(String::new("default"))
-}
-
-pub(crate) fn generate_slot_outlet_name(ctx: &mut CodegenContext, el: &ElementNode<'_>) {
-    match get_slot_outlet_name(el) {
-        SlotOutletName::Static(name) => {
-            ctx.push("\"");
-            ctx.push(&escape_js_string(name.as_str()));
-            ctx.push("\"");
-        }
-        SlotOutletName::Dynamic(exp) => generate_expression(ctx, exp),
-    }
-}
-
-pub(crate) fn has_slot_outlet_props(el: &ElementNode<'_>) -> bool {
-    el.props.iter().any(slot_outlet_prop_generates_output)
-}
-
-pub(crate) fn generate_slot_outlet_props_entries(ctx: &mut CodegenContext, el: &ElementNode<'_>) {
-    let static_merge = crate::codegen::props::StaticMerge::from_props(&el.props);
-
-    let has_dynamic_class = el.props.iter().any(|prop| match prop {
-        PropNode::Directive(dir) if !is_slot_name_bind(dir) && dir.name.as_str() == "bind" => {
-            matches!(
-                dir.arg.as_ref(),
-                Some(ExpressionNode::Simple(exp))
-                    if exp.is_static && exp.content.as_str() == "class"
-            )
-        }
-        _ => false,
-    });
-
-    let has_dynamic_style = el.props.iter().any(|prop| match prop {
-        PropNode::Directive(dir) if !is_slot_name_bind(dir) && dir.name.as_str() == "bind" => {
-            matches!(
-                dir.arg.as_ref(),
-                Some(ExpressionNode::Simple(exp))
-                    if exp.is_static && exp.content.as_str() == "style"
-            )
-        }
-        _ => false,
-    });
-
-    let mut first = true;
-    for prop in &el.props {
-        if is_slot_name_prop(prop) {
-            continue;
-        }
-
-        match prop {
-            PropNode::Attribute(attr) => {
-                if (attr.name.as_str() == "class" && has_dynamic_class)
-                    || (attr.name.as_str() == "style" && has_dynamic_style)
-                {
-                    continue;
-                }
-
-                if !first {
-                    ctx.push(", ");
-                }
-
-                let key = camelize(&attr.name);
-                if is_valid_js_identifier(&key) {
-                    ctx.push(&key);
-                } else {
-                    ctx.push("\"");
-                    ctx.push(&escape_js_string(&key));
-                    ctx.push("\"");
-                }
-                ctx.push(": ");
-                if let Some(value) = &attr.value {
-                    ctx.push("\"");
-                    ctx.push(&escape_js_string(value.content.as_str()));
-                    ctx.push("\"");
-                } else {
-                    ctx.push("\"\"");
-                }
-                first = false;
-            }
-            PropNode::Directive(dir) => {
-                if !is_supported_directive(dir)
-                    || (dir.arg.is_none()
-                        && (dir.name.as_str() == "bind" || dir.name.as_str() == "on"))
-                {
-                    continue;
-                }
-
-                if !first {
-                    ctx.push(", ");
-                }
-                generate_slot_outlet_directive_prop_with_static(ctx, dir, static_merge);
-                first = false;
-            }
-        }
-    }
-}
-
-pub(crate) fn generate_slot_outlet_props(ctx: &mut CodegenContext, el: &ElementNode<'_>) {
-    generate_slot_outlet_props_inner(ctx, el, None);
-}
-
-pub(crate) fn generate_slot_outlet_props_with_key(
-    ctx: &mut CodegenContext,
-    el: &ElementNode<'_>,
-    generate_key: &dyn Fn(&mut CodegenContext),
-) {
-    generate_slot_outlet_props_inner(ctx, el, Some(generate_key));
-}
-
-fn generate_slot_outlet_props_inner(
-    ctx: &mut CodegenContext,
-    el: &ElementNode<'_>,
-    generate_key: Option<&dyn Fn(&mut CodegenContext)>,
-) {
-    let has_vbind_object = has_slot_outlet_vbind_object(el);
-    let has_von_object = has_slot_outlet_von_object(el);
-    let has_entries = has_slot_outlet_entry_props(el);
-    let has_key = generate_key.is_some();
-
-    if has_vbind_object || has_von_object {
-        let needs_merge = has_key || has_entries || (has_vbind_object && has_von_object);
-
-        if needs_merge {
-            ctx.use_helper(RuntimeHelper::MergeProps);
-            ctx.push(ctx.helper(RuntimeHelper::MergeProps));
-            ctx.push("(");
-            let mut first = true;
-
-            if has_vbind_object {
-                generate_vbind_object_exp(ctx, &el.props);
-                first = false;
-            }
-
-            if has_von_object {
-                if !first {
-                    ctx.push(", ");
-                }
-                generate_von_object_exp(ctx, &el.props);
-                first = false;
-            }
-
-            if has_key || has_entries {
-                if !first {
-                    ctx.push(", ");
-                }
-                generate_slot_outlet_props_object(ctx, el, generate_key, has_entries);
-            }
-
-            ctx.push(")");
-        } else if has_vbind_object {
-            ctx.use_helper(RuntimeHelper::NormalizeProps);
-            ctx.use_helper(RuntimeHelper::GuardReactiveProps);
-            ctx.push(ctx.helper(RuntimeHelper::NormalizeProps));
-            ctx.push("(");
-            ctx.push(ctx.helper(RuntimeHelper::GuardReactiveProps));
-            ctx.push("(");
-            generate_vbind_object_exp(ctx, &el.props);
-            ctx.push("))");
-        } else {
-            generate_von_object_exp(ctx, &el.props);
-        }
-        return;
-    }
-
-    generate_slot_outlet_props_object(ctx, el, generate_key, has_entries);
-}
-
-fn generate_slot_outlet_props_object(
-    ctx: &mut CodegenContext,
-    el: &ElementNode<'_>,
-    generate_key: Option<&dyn Fn(&mut CodegenContext)>,
-    has_entries: bool,
-) {
-    ctx.push("{");
-    let mut needs_separator = false;
-
-    if let Some(generate_key) = generate_key {
-        ctx.push(" key: ");
-        generate_key(ctx);
-        needs_separator = true;
-    }
-
-    if has_entries {
-        if needs_separator {
-            ctx.push(", ");
-        } else {
-            ctx.push(" ");
-        }
-        generate_slot_outlet_props_entries(ctx, el);
-        needs_separator = true;
-    }
-
-    if needs_separator {
-        ctx.push(" ");
-    }
-    ctx.push("}");
-}
-
-/// Check if component has slot children that need to be generated as slots object
-pub fn has_slot_children(el: &ElementNode<'_>) -> bool {
-    if el.children.is_empty() {
-        return false;
-    }
-
-    // Teleport and KeepAlive consume raw children rather than a slot object.
-    // KeepAlive still gets DYNAMIC_SLOTS at the vnode patch-flag layer.
-    if matches!(
-        el.tag.as_str(),
-        "Teleport" | "teleport" | "KeepAlive" | "keep-alive"
-    ) {
-        return false;
-    }
-
-    // Check for v-slot on component root
-    for prop in &el.props {
-        if let PropNode::Directive(dir) = prop
-            && dir.name.as_str() == "slot"
-        {
-            return true;
-        }
-    }
-
-    // If children consist only of whitespace text and/or comments, skip slot generation.
-    // This matches Vue's official compiler behavior where `<Comp> </Comp>` does not
-    // produce a default slot (important for <router-view>, <transition>, etc.).
-    let has_meaningful_child = el.children.iter().any(|child| match child {
-        TemplateChildNode::Text(t) => !t.content.trim().is_empty(),
-        TemplateChildNode::Comment(_) => false,
-        _ => true,
-    });
-    if !has_meaningful_child {
-        return false;
-    }
-
-    // Check for any children (default slot) or template slots
-    true
-}
-
-/// Check if component has dynamic slots (requires DYNAMIC_SLOTS patch flag)
-pub fn has_dynamic_slots_flag(el: &ElementNode<'_>) -> bool {
-    let collected_slots = collect_slots(el);
-    if collected_slots.iter().any(|s| s.is_dynamic) {
-        return true;
-    }
-    if has_forwarded_slot_outlet(el) {
-        return true;
-    }
-    // Also check for v-if/v-for on slot templates (they become IfNode/ForNode children)
-    has_conditional_or_loop_slots(el)
-}
-
-/// Check whether this component forwards an incoming slot to another component,
-/// e.g. `<Inner><slot /></Inner>`.
-fn has_forwarded_slot_outlet(el: &ElementNode<'_>) -> bool {
-    el.children.iter().any(child_contains_slot_outlet)
-}
-
-fn child_contains_slot_outlet(child: &TemplateChildNode<'_>) -> bool {
-    match child {
-        TemplateChildNode::Element(el) => {
-            if el.tag_type == ElementType::Slot || el.tag.as_str() == "slot" {
-                return true;
-            }
-            el.children.iter().any(child_contains_slot_outlet)
-        }
-        TemplateChildNode::If(if_node) => if_node
-            .branches
-            .iter()
-            .flat_map(|branch| branch.children.iter())
-            .any(child_contains_slot_outlet),
-        TemplateChildNode::For(for_node) => {
-            for_node.children.iter().any(child_contains_slot_outlet)
-        }
-        _ => false,
-    }
-}
-
-/// Check if children have conditional (v-if) or looped (v-for) slot templates.
-/// Only returns true when the IfNode/ForNode wraps a `<template v-slot>` element.
-fn has_conditional_or_loop_slots(el: &ElementNode<'_>) -> bool {
-    el.children.iter().any(|child| match child {
-        TemplateChildNode::If(if_node) => if_node.branches.iter().any(|branch| {
-            branch.children.iter().any(|c| {
-                if let TemplateChildNode::Element(el) = c {
-                    el.tag.as_str() == "template" && has_v_slot(el)
-                } else {
-                    false
-                }
-            })
-        }),
-        TemplateChildNode::For(for_node) => for_node.children.iter().any(|c| {
-            if let TemplateChildNode::Element(el) = c {
-                el.tag.as_str() == "template" && has_v_slot(el)
-            } else {
-                false
-            }
-        }),
-        _ => false,
-    })
-}
-
-fn child_is_slot_template(child: &TemplateChildNode<'_>) -> bool {
-    match child {
-        TemplateChildNode::Element(el) => el.tag.as_str() == "template" && has_v_slot(el),
-        TemplateChildNode::If(if_node) => if_node.branches.iter().any(|branch| {
-            branch.children.iter().any(|child| {
-                matches!(
-                    child,
-                    TemplateChildNode::Element(el)
-                        if el.tag.as_str() == "template" && has_v_slot(el)
-                )
-            })
-        }),
-        TemplateChildNode::For(for_node) => for_node.children.iter().any(|child| {
-            matches!(
-                child,
-                TemplateChildNode::Element(el)
-                    if el.tag.as_str() == "template" && has_v_slot(el)
-            )
-        }),
-        _ => false,
-    }
-}
-
-fn slot_children_have_meaningful_content(children: &[&TemplateChildNode<'_>]) -> bool {
-    children.iter().any(|child| match child {
-        TemplateChildNode::Text(text) => !text.content.trim().is_empty(),
-        TemplateChildNode::Comment(_) => false,
-        _ => true,
-    })
-}
+use super::super::context::CodegenContext;
+use super::super::expression::generate_expression;
+use super::super::helpers::{escape_js_string, is_valid_js_identifier};
+use super::super::node::generate_node;
+use super::detect::{
+    child_is_slot_template, has_conditional_or_loop_slots, has_forwarded_slot_outlet,
+    slot_children_have_meaningful_content,
+};
+use super::params::{extract_slot_params, get_slot_props, prefix_slot_defaults};
 
 /// Generate slots object for component
 pub fn generate_slots(ctx: &mut CodegenContext, el: &ElementNode<'_>) {
@@ -876,17 +378,17 @@ fn generate_looped_slot(ctx: &mut CodegenContext, for_node: &ForNode<'_>) {
 
     if let Some(value) = &for_node.value_alias {
         generate_expression(ctx, value);
-        super::v_for::helpers::extract_for_params(value, &mut callback_params);
+        super::super::v_for::helpers::extract_for_params(value, &mut callback_params);
     }
     if let Some(key) = &for_node.key_alias {
         ctx.push(", ");
         generate_expression(ctx, key);
-        super::v_for::helpers::extract_for_params(key, &mut callback_params);
+        super::super::v_for::helpers::extract_for_params(key, &mut callback_params);
     }
     if let Some(index) = &for_node.object_index_alias {
         ctx.push(", ");
         generate_expression(ctx, index);
-        super::v_for::helpers::extract_for_params(index, &mut callback_params);
+        super::super::v_for::helpers::extract_for_params(index, &mut callback_params);
     }
 
     ctx.add_slot_params(&callback_params);
@@ -1022,7 +524,7 @@ fn generate_slot_children(ctx: &mut CodegenContext, children: &[TemplateChildNod
             match child {
                 TemplateChildNode::Text(text) => {
                     ctx.push("\"");
-                    ctx.push(&super::helpers::escape_js_string(&text.content));
+                    ctx.push(&super::super::helpers::escape_js_string(&text.content));
                     ctx.push("\"");
                 }
                 TemplateChildNode::Interpolation(interp) => {
@@ -1059,7 +561,7 @@ fn generate_slot_child_node(ctx: &mut CodegenContext, child: &TemplateChildNode<
             ctx.use_helper(RuntimeHelper::CreateText);
             ctx.push(ctx.helper(RuntimeHelper::CreateText));
             ctx.push("(\"");
-            ctx.push(&super::helpers::escape_js_string(&text.content));
+            ctx.push(&super::super::helpers::escape_js_string(&text.content));
             ctx.push("\")");
         }
         TemplateChildNode::Interpolation(interp) => {
@@ -1131,67 +633,4 @@ fn strip_ctx_prefix_for_slot_params(ctx: &CodegenContext, content: &str) -> Stri
         result = String::from(replaced);
     }
     result
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{is_valid_js_identifier, prefix_slot_defaults};
-
-    #[test]
-    fn test_is_valid_js_identifier_valid() {
-        assert!(is_valid_js_identifier("foo"));
-        assert!(is_valid_js_identifier("_bar"));
-        assert!(is_valid_js_identifier("$baz"));
-        assert!(is_valid_js_identifier("foo123"));
-        assert!(is_valid_js_identifier("camelCase"));
-        assert!(is_valid_js_identifier("PascalCase"));
-    }
-
-    #[test]
-    fn test_is_valid_js_identifier_invalid() {
-        assert!(!is_valid_js_identifier("123foo")); // starts with number
-        assert!(!is_valid_js_identifier("")); // empty
-        assert!(!is_valid_js_identifier("foo-bar")); // contains hyphen
-        assert!(!is_valid_js_identifier("foo.bar")); // contains dot
-        assert!(!is_valid_js_identifier("foo bar")); // contains space
-        assert!(!is_valid_js_identifier("item-header")); // hyphenated slot name
-    }
-
-    #[test]
-    fn test_hyphenated_slot_names_need_quotes() {
-        assert!(!is_valid_js_identifier("item-header"));
-        assert!(!is_valid_js_identifier("card-body"));
-        assert!(!is_valid_js_identifier("main-content"));
-        assert!(!is_valid_js_identifier("list-item"));
-    }
-
-    #[test]
-    fn test_regular_slot_names_are_valid_identifiers() {
-        assert!(is_valid_js_identifier("default"));
-        assert!(is_valid_js_identifier("header"));
-        assert!(is_valid_js_identifier("footer"));
-        assert!(is_valid_js_identifier("content"));
-    }
-
-    #[test]
-    fn test_prefix_slot_defaults() {
-        // Default values should get _ctx. prefix
-        assert_eq!(
-            prefix_slot_defaults("{ item = defaultItem }"),
-            "{ item = _ctx.defaultItem }"
-        );
-        assert_eq!(prefix_slot_defaults("{ count = 0 }"), "{ count = 0 }");
-        assert_eq!(
-            prefix_slot_defaults("{ name = 'test' }"),
-            "{ name = 'test' }"
-        );
-        // Literals should not be prefixed
-        assert_eq!(prefix_slot_defaults("{ x = true }"), "{ x = true }");
-        assert_eq!(prefix_slot_defaults("{ x = false }"), "{ x = false }");
-        assert_eq!(prefix_slot_defaults("{ x = null }"), "{ x = null }");
-        assert_eq!(
-            prefix_slot_defaults("{ x = undefined }"),
-            "{ x = undefined }"
-        );
-    }
 }
